feat: improve layout with breakpoints
