+++ src/pages/App/styled.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const AppContainer = styled.div`
+  max-width: var(--screen-max-width);
+  margin: 0 auto;
+`;
